available before version than go1.16
